user page 403

+++ app/Http/Controllers/RoleController.php
@@ -7,7 +7,7 @@ use App\Models\Permission;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
-use App\Http\Requests\RequestIndexRoles;
+use Illuminate\Support\Facades\Session;
 use App\Http\Requests\RequestStoreRoles;
 use Illuminate\Support\Facades\Response;
 use App\Http\Requests\RequestUpdateRoles;
@@ -16,11 +16,11 @@ class RoleController extends Controller
 {
     public function index()
     {        
-        $roles = Role::paginate(15);
-        // if(!Auth::user()->hasRole('Super') || !Auth::user()->hasRole($role->name)){
-        //     return view('Admin.error.403');
-        // }                 
+        if(!Auth::user()->can(['grupo.criar', 'grupo.editar','grupo.visualizar', 'grupo.remover'])){
+            return view('Admin.error.403');
+        }    
 
+        $roles = Role::paginate(15);
         $permissions = Permission::join('role_has_permissions', 'permissions.id', 'role_has_permissions.permission_id')
         ->groupBy('permissions.name')
         ->select('permissions.name')
@@ -33,6 +33,10 @@ class RoleController extends Controller
     }
     public function create()
     {
+        if(!Auth::user()->can(['grupo.criar', 'grupo.editar','grupo.visualizar', 'grupo.remover'])){
+            return view('Admin.error.403');
+        } 
+
         $permissions = Permission::all();
         return view('Admin.cruds.group.create', [
             'permissions'=>$permissions
@@ -44,12 +48,16 @@ class RoleController extends Controller
             'name' => $request->name
         ]);
         $role->syncPermissions($request->permissions);
+        Session::flash('success','Grupo cadastrado com sucesso!');
         return redirect()->route('admin.dashboard.group.index');
     }
     public function edit(Request $request,Role $role)
     { 
-        $permissions = Permission::all();
+        if(!Auth::user()->can(['grupo.criar', 'grupo.editar','grupo.visualizar', 'grupo.remover'])){
+            return view('Admin.error.403');
+        } 
 
+        $permissions = Permission::all();
         return view('Admin.cruds.group.edit', [
             'role'=>$role,
             'permissions'=>$permissions
@@ -72,6 +80,7 @@ class RoleController extends Controller
             ]);
             $role->syncPermissions($request->permissions);
             DB::commit();
+            Session::flash('success','Grupo alterado com sucesso!');
             return redirect()->route('admin.dashboard.group.index');
         }catch (\Exception $exception){
             DB::rollBack();
@@ -81,19 +90,28 @@ class RoleController extends Controller
 
     public function destroy(Request $request,Role $role)
     {
+        if(!Auth::user()->can('grupo.remover')){
+            return view('Admin.error.403');
+        } 
+
         $role->delete();
+        Session::flash('success','Grupo deletado com sucesso!');
         return redirect()->back();
     }
     public function destroySelected(Request $request)
     {
         if($deleted = Role::whereIn('id', $request->deleteAll)->delete()){
             
-            return Response::json(['status' => 'success', 'message' => $deleted.' itens deletados com sucessso']);
+            return Response::json(['status' => 'success', 'message' => $deleted.' itens deletados com sucessso!']);
         }
     }
 
     public function sorting(Request $request)
-    {
+    {   
+        if(!Auth::user()->can('grupo.remover')){
+            return view('Admin.error.403');
+        }
+
         foreach($request->arrId as $sorting => $id){
             Role::where('id', $id)->update(['sorting' => $sorting]);
         }
